requirements.txt
# Core Framework
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Scientific Computing (Python 3.13 호환)
numpy>=1.26.0

# AI APIs
openai>=1.3.0
anthropic>=0.7.0

# PyTorch (CPU 버전)
torch>=2.1.0
torchvision>=0.16.0

# Data Processing & Search
ddgs>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
scikit-learn>=1.4.0

# Utilities
pyyaml>=6.0.1

# Build Tools (Python 3.13 필수)
setuptools>=68.0.0
wheel>=0.41.0
